+++ QLMarkdown/AboutViewController.swift
@@ -56,7 +56,7 @@ class AboutViewController: NSViewController {
         s += "\(String(cString: get_lua_info()))<br />\n"
         s += "Enry (<a href=\"https://www.github.com/go-enry/go-enry/\">https://www.github.com/go-enry/go-enry/</a>)<br />\n"
         s += "<br />\n———<br />\n<br />\n"
-        s += "Thanks to hazarek (<a href=\"https://github.com/hazarek\">https://github.com/hazarek</a>) for the css theme.<br />\n"
+        s += "Thanks to hazarek (<a href=\"https://github.com/hazarek\">https://github.com/hazarek</a>) for icon and the css theme.<br />\n"
         
         s += "</div>"